+++ src/features/kmap/KMapGenerator.css
