+++ src/features/kmap/components/InputControls.jsx
@@ -1,4 +1,4 @@
-import '../../../shared/styles/KMapGenerator.css';
+import '../KMapGenerator.css';
 import React from 'react';
 import {
     Cpu,
