+++ src/features/kmap/components/TruthTableDisplay.jsx
@@ -1,4 +1,4 @@
-import '../../../shared/styles/KMapGenerator.css';
+import '../KMapGenerator.css';
 import React, { useState, memo } from "react";
 import { Maximize2, X } from 'lucide-react';
 import { evaluate } from 'mathjs';
