+++ src/features/kmap/hooks/index.js
@@ -1,0 +1,2 @@
+export { default as useKMapLogic } from './useKMapLogic';
+export { default as useSpeechSynthesis } from './useSpeechSynthesis';
