+++ src/features/kmap/hooks/useKMapLogic.js
@@ -124,3 +124,5 @@ export const useKMapLogic = (numVariables, variables, inputValue, dontCares = ''
         optimizationType
     };
 };
+
+export default useKMapLogic;
