+++ src/features/kmap/hooks/useSpeechSynthesis.js
@@ -84,3 +84,5 @@ export const useSpeechSynthesis = () => {
         voices
     };
 };
+
+export default useSpeechSynthesis;
