+++ src/features/kmap/utils/GroupDetector.js
