+++ src/features/kmap/utils/index.js
@@ -1,0 +1,2 @@
+export { default as QuineMcCluskey } from './QuineMcCluskey';
+export { default as detectGroups } from './GroupDetector';
